# Improve performance by ~30% by reducing calls to win32 api