## Summary
* `IsInsideDirectory: False`: 27.3% faster: 79.6ms --> 57.84ms
* `IsInsideDirectory: True `:  34.5% faster: 208.01ms --> 136.17ms

## Before (1.0.3)

|                               Method | NumFiles | FileSize | IsInsideDirectory |      Mean |    Error |   StdDev |
|------------------------------------- |--------- |--------- |------------------ |----------:|---------:|---------:|
| **ForceOps.Lib.FileAndDirectoryDeleter** |     **1000** |       **10** |             **False** |  **79.60 ms** | **1.573 ms** | **3.485 ms** |
|           System.IO.Directory.Delete |     1000 |       10 |             False |  60.59 ms | 1.205 ms | 2.569 ms |
| **ForceOps.Lib.FileAndDirectoryDeleter** |     **1000** |       **10** |              **True** | **208.01 ms** | **4.111 ms** | **6.400 ms** |
|           System.IO.Directory.Delete |     1000 |       10 |              True | 139.70 ms | 2.541 ms | 3.392 ms |

## After

|                               Method | NumFiles | FileSize | IsInsideDirectory |      Mean |    Error |   StdDev |
|------------------------------------- |--------- |--------- |------------------ |----------:|---------:|---------:|
| **ForceOps.Lib.FileAndDirectoryDeleter** |     **1000** |       **10** |             **False** |  **57.84 ms** | **1.009 ms** | **1.276 ms** |
|           System.IO.Directory.Delete |     1000 |       10 |             False |  57.38 ms | 1.106 ms | 1.398 ms |
| **ForceOps.Lib.FileAndDirectoryDeleter** |     **1000** |       **10** |              **True** | **136.20 ms** | **2.203 ms** | **1.953 ms** |
|           System.IO.Directory.Delete |     1000 |       10 |              True | 136.17 ms | 2.650 ms | 2.349 ms |